Satisfy test suite with new XS based config

+++ xs/src/Point.hpp
@@ -47,9 +47,9 @@ class Point
 class Pointf
 {
     public:
-    float x;
-    float y;
-    explicit Pointf(float _x = 0, float _y = 0): x(_x), y(_y) {};
+    double x;
+    double y;
+    explicit Pointf(double _x = 0, double _y = 0): x(_x), y(_y) {};
     
     #ifdef SLIC3RXS
     void from_SV(SV* point_sv);
